Enable incremental build and remove Mono
Instead of running the package install script after the build of each of
the project, it is run only after the build of Peachpie.Compiler.Tools.
The script was also modified so that it won't reinstall the packages if
it doesn't find Mono, which was needed to run nuget.exe outside Windows.
Therefore, we don't have to install it in Travis CI.